migrate execdetails inside the stmtctx and related methods to execdetails
