+++ pkg/util/stmtsummary/statement_summary.go
@@ -238,7 +238,7 @@ type StmtExecInfo struct {
 	ParseLatency        time.Duration
 	CompileLatency      time.Duration
 	StmtCtx             *stmtctx.StatementContext
-	CopTasks            *stmtctx.CopTasksDetails
+	CopTasks            *execdetails.CopTasksDetails
 	ExecDetail          *execdetails.ExecDetails
 	MemMax              int64
 	DiskMax             int64
